Started implementation of the Waiting Screen. This is the screen the users wait for the game to start.
The screen builder listens to changes in the GameDoc on firebase, so it will always present the latest data about the game.

The screen should be somewhat different between the owner of the game and other players.
We haven't addressed this matter yet. ( And in general the UI in this screen is lacking for now).